Change the script for the bootship quest
