+++ quests/story/bootship.questtemplate.patch
@@ -1,0 +1,7 @@
+[
+	{
+		"op" : "replace",
+		"path" : "/script",
+		"value" : "/quests/scripts/story/fu_bootship.lua"
+	}
+]
